move to private the submodules
