+++ netket/utils/group/_semigroup.py
